userns: device cgroup guard to allow mknod in non-initial userns
If a container manager restricts its unprivileged (user namespaced)
children by a device cgroup, it is not necessary to deny mknod
anymore. Thus, user space applications may map devices on different
locations in the file system by using mknod() inside the container.

A use case for this, we also use in GyroidOS, is to run virsh for
VMs inside an unprivileged container. virsh creates device nodes,
e.g., "/var/run/libvirt/qemu/11-fgfg.dev/null" which currently fails
in a non-initial userns, even if a cgroup device white list with the
corresponding major, minor of /dev/null exists. Thus, in this case
the usual bind mounts or pre populated device nodes under /dev are
not sufficient.

To circumvent this limitation, we allow mknod() if device cgroup
restriction applies to the current task and check CAP_MKNOD for
the current user namespace by ns_capable() instead of the global
CAP_MKNOD by capable(). Further, the usual default for non-initial
user namespace mounts SB_I_NODEV is not set in case the device
cgroup guard is active.

To avoid user space regressions, this feature is bound to a flag
inside the user namespace. Similarly to the 'setgroups' knob, we
implemented the 'devcg_guard' knob based on the procfs file
'/proc/<pid>/devcg_guard'. "allow" activates the devcg guard and
"deny" deactivates it. The default is "deny". Writing "allow" to
'/proc/<pid>/devcg_guard' only succeeds if the caller has CAP_MKNOD
over the parent user namespace. Otherwise and in the initial user
namespace, -EPERM is returned.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>