security: provide unified security_device_access() hook
Provide a new lsm hook which may be used to check permission on
a device by its dev_t representation only. This hook will be used
to unify all devcgroup_* permission hooks in one place and let
other LSMs implement device access mangment on basis of this new
hook.

Let device_cgroup implement this as lsm hook. But remain backward
compatiblitiy if CONFIG_SECURITY is not set similar to commoncap.

Signed-off-by: Michael Weiß <michael.weiss@aisec.fraunhofer.de>